added some scripts in application.html, installed bootsrap, not many changes
