+++ app/assets/stylesheets/application.css.scss
@@ -13,3 +13,6 @@
  *= require_tree .
  *= require_self
  */
+
+@import "bootstrap-sprockets";
+@import "bootstrap";
