+++ app/assets/stylesheets/dogs.scss
