+++ app/views/auth/new.html.erb
@@ -1,4 +1,6 @@
+
 <h1>Login</h1>
+
 <%= flash[:message] %>
 <%= form_for :auth do |f| %>
     <%= f.label :username %>
