+++ app/views/chats/show.html.erb
@@ -1,1 +1,1 @@
-
+<h1>Chat Message</h1>
