+++ app/views/users/show.html.erb
@@ -17,7 +17,6 @@ Occupation: <%= @user.occupation %><br>
 </ul>
 
 <%= button_to "Add Dog", new_dog_path, method: :get %>
-
 <%= button_to "Logout", logout_path, method: :delete %>
 <%= button_to "Delete Account", @user, method: :delete %>
