+++ config/routes.rb
@@ -8,6 +8,6 @@ Rails.application.routes.draw do
   get '/login', to: 'auth#new'
   post '/login', to: 'auth#create'
   delete '/login', to: 'auth#destroy', as: 'logout'
-
+  # root to: "users#index"
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
